fix null mx detection
Net::DNS returns the host name for Null MX records as the string '.'. As
a result, Null MX records are never detected by version 1.203 of
Email::Valid.

valid.t runs a check against nomail.digilicious.com, which (currently)
has no A record. For the test suite to verify that the Null MX check
works correctly, it needs to check a name which has both a Null MX and
an A record. Otherwise, the missing A record will always lead to the
address check failing (and the unit test therefore succeeding), with the
missing A record masking the result of the Null MX detection.

A DNS name that currently has a Null MX and address records would be
ajnn.de. However, there is no guarantee that it’ll stay that way, so
this particular name is at most useful for development testing, not for
automated testing.

Since I don’t know which DNS name I should use in the test suite, I
haven’t provided a PR to fix this issue. The fix itself is very simple
though.